requirements.txt
eth-brownie==1.17.2
Faker
